gas-optimize deposit
update storage mapping for major use-case (handleOps)

+++ contracts/core/EntryPoint.sol
@@ -447,7 +447,7 @@ contract EntryPoint is IEntryPoint, StakeManager, NonceManager, ReentrancyGuard,
                 if (requiredPrefund > deposit) {
                     revert FailedOp(opIndex, "AA21 didn't pay prefund");
                 }
-                senderInfo.deposit = uint112(deposit - requiredPrefund);
+                senderInfo.deposit = deposit - requiredPrefund;
             }
             gasUsedByValidateAccountPrepayment = preGas - gasleft();
         }
@@ -488,7 +488,7 @@ contract EntryPoint is IEntryPoint, StakeManager, NonceManager, ReentrancyGuard,
             if (deposit < requiredPreFund) {
                 revert FailedOp(opIndex, "AA31 paymaster deposit too low");
             }
-            paymasterInfo.deposit = uint112(deposit - requiredPreFund);
+            paymasterInfo.deposit = deposit - requiredPreFund;
             try
                 IPaymaster(paymaster).validatePaymasterUserOp{gas: gas}(
                     op,
